+++ packages/gatsby/src/redux/index.js
@@ -46,6 +46,15 @@ try {
   }
   if (initialState.nodes) {
     initialState.nodes = objectToMap(initialState.nodes)
+
+    initialState.nodesByType = new Map()
+    initialState.nodes.forEach(node => {
+      const { type } = node.internal
+      if (!initialState.nodesByType.has(type)) {
+        initialState.nodesByType.set(type, new Map())
+      }
+      initialState.nodesByType.get(type).set(node.id, node)
+    })
   }
 } catch (e) {
   // ignore errors.
@@ -63,7 +72,8 @@ const store = Redux.createStore(
 )
 
 // Persist state.
-const saveState = state => {
+function saveState() {
+  const state = store.getState()
   const pickedState = _.pick(state, [
     `nodes`,
     `status`,
@@ -85,33 +95,13 @@ const saveState = state => {
     () => {}
   )
 }
-const saveStateDebounced = _.debounce(saveState, 1000)
+
+exports.saveState = saveState
 
 store.subscribe(() => {
   const lastAction = store.getState().lastAction
   emitter.emit(lastAction.type, lastAction)
 })
-
-// During development, once bootstrap is finished, persist state on changes.
-let bootstrapFinished = false
-if (process.env.gatsby_executing_command === `develop`) {
-  emitter.on(`BOOTSTRAP_FINISHED`, () => {
-    bootstrapFinished = true
-    saveState(store.getState())
-  })
-  emitter.on(`*`, () => {
-    if (bootstrapFinished) {
-      saveStateDebounced(store.getState())
-    }
-  })
-}
-
-// During builds, persist state once bootstrap has finished.
-if (process.env.gatsby_executing_command === `build`) {
-  emitter.on(`BOOTSTRAP_FINISHED`, () => {
-    saveState(store.getState())
-  })
-}
 
 /** Event emitter */
 exports.emitter = emitter