+++ packages/gatsby/src/redux/nodes.js
@@ -16,17 +16,40 @@ const getNodes = () => {
 
 exports.getNodes = getNodes
 
-const getNode = id => store.getState().nodes.get(id)
-
 /** Get node by id from store.
  *
  * @param {string} id
  * @returns {Object}
  */
+const getNode = id => store.getState().nodes.get(id)
+
 exports.getNode = getNode
 
-exports.getNodesByType = type =>
-  getNodes().filter(node => node.internal.type === type)
+/**
+ * Get all nodes of type from redux store.
+ *
+ * @param {string} type
+ * @returns {Array}
+ */
+const getNodesByType = type => {
+  const nodes = store.getState().nodesByType.get(type)
+  if (nodes) {
+    return Array.from(nodes.values())
+  } else {
+    return []
+  }
+}
+
+exports.getNodesByType = getNodesByType
+
+/**
+ * Get all types from redux store.
+ *
+ * @returns {Array}
+ */
+const getTypes = () => Array.from(store.getState().nodesByType.keys())
+
+exports.getTypes = getTypes
 
 /**
  * Determine if node has changed.