+++ www/src/views/starter-library/index.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react"
-import Helmet from "react-helmet"
+import { Helmet } from "react-helmet"
 import Layout from "../../components/layout"
+import Unbird from "../../components/unbird"
 import RRSM from "../../utils/reach-router-state-manager"
 import queryString from "query-string"
 
@@ -19,8 +20,8 @@ class StarterLibraryPage extends Component {
       urlState.s !== undefined
         ? urlState.s // if theres a search term
         : urlState.d && !Array.isArray(urlState.d)
-          ? urlState.d // if theres a single dependency
-          : `Library` // if no search term or single dependency
+        ? urlState.d // if theres a single dependency
+        : `Library` // if no search term or single dependency
 
     return (
       <Layout location={location}>
@@ -54,6 +55,11 @@ class StarterLibraryPage extends Component {
             />
           )}
           defaultSearchState={{ v: [`2`] }}
+        />
+        <Unbird
+          dataSetId="5c113a828240aa564734d954"
+          publicKey={process.env.GATSBY_FEEDBACK_KEY_STARTERLIB}
+          feedbackPrompt="Have feedback on the Starter Library?"
         />
       </Layout>
     )