+++ packages/gatsby-plugin-typescript/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.7...gatsby-plugin-typescript@2.0.8) (2019-02-13)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.0.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.6...gatsby-plugin-typescript@2.0.7) (2019-02-04)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.0.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.5...gatsby-plugin-typescript@2.0.6) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.4...gatsby-plugin-typescript@2.0.5) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.0.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.3...gatsby-plugin-typescript@2.0.4) (2019-01-28)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+<a name="2.0.3"></a>
+
+## [2.0.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.2...gatsby-plugin-typescript@2.0.3) (2018-12-05)
+
+### Features
+
+- **gatsby-plugin-typescript:** allow specifying babel preset options ([#10248](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/issues/10248)) ([f106241](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/commit/f106241))
+
 <a name="2.0.2"></a>
 
 ## [2.0.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.1...gatsby-plugin-typescript@2.0.2) (2018-11-29)